Add basic InitialScreen tests
